% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{availableDatasets}
\alias{availableDatasets}
\alias{availableDatasets_by_source}
\alias{availableDataSources}
\title{Available Datasets}
\usage{
availableDatasets(pagelimit = 1e+06)

availableDatasets_by_source(id)

availableDataSources()
}
\arguments{
\item{id}{Numeric or String Numeric. An identifier of the organisation/authority providing the dataset}
}
\value{
Returns a data frame
}
\description{
Obtains the list of available datasets from the \url{data.gov.uz}
}
\section{Functions}{
\itemize{
\item \code{availableDatasets_by_source()}: Obtains the list of available datasets published by particular organisation/authority

\item \code{availableDataSources()}: Obtains the list of available datasources - organisations

}}
\examples{
\dontrun{
 availableDatasets()
 availableDataSources()
 availableDatasets_by_source(143) }
}
\seealso{
\code{\link{getData}} to download available datasets by their \code{id}
}
\author{
Alisher Suyunov
}
