% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FX.R
\name{get_FX}
\alias{get_FX}
\title{Downloading Foreign Currency Exchange Data}
\usage{
get_FX(currency = c("USD", "EUR"), from = "01-10-2020", to = "dd-mm-YYYY")
}
\arguments{
\item{currency}{A currency code according to ISO 4217 (e.g. USD or EUR)}

\item{from}{A string representing the start date of the period of interest in date-month-year date format}

\item{to}{A string representing the end date of the period of interest in date-month-year date format}
}
\value{
Returns a data frame
}
\description{
Obtains foreign currency exchange rate from the Republican Currency Exchange over the period given
}
\examples{
 \dontrun{
 get_FX("USD")
 get_FX("EUR", from = "01-01-2018", to = "12-10-2020")}
}
\seealso{
\code{\link{getSecurities}} or \code{\link{RegisteredSecurities}} to see the complete list of available securities

\code{\link{getTicker}} to download stock market prices

\code{\link{getMarketIndex}} to download stock market indices by sector
}
\author{
Alisher Suyunov
}
