% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllIssuers.R
\name{getAllIssuers}
\alias{getAllIssuers}
\title{Downloads a full list of issuers}
\usage{
getAllIssuers()
}
\value{
Returns a data frame with the list of issuers
}
\description{
Returns a data frame with the list of issuers from the Central Securities Depository (\url{uzcsd.uz}) database
}
\examples{
 \dontrun{
 getAllIssuers()}
}
\author{
Alisher Suyunov
}
