% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\alias{getData_dictionary}
\alias{getData_history}
\title{Downloading Datasets}
\usage{
getData(id, format = "csv", sep = ";")

getData_dictionary(id, sep = ";", header = FALSE)

getData_history(id)
}
\arguments{
\item{id}{a numeric or character that represents the dataset identifier}

\item{format}{a character. A format (CSV or JSON) of the response from the server}

\item{sep}{a field separator character. Values on each line of the file are separated by this character. By default it is a semicolon ';'. The separator character is passed to the function read.csv.}

\item{header}{a logical value representing if the file contains variable names on its first line}
}
\value{
Returns a data frame
}
\description{
Downloads available datasets and related information from data.gov.uz API
}
\section{Functions}{
\itemize{
\item \code{getData_dictionary}: Downloads a dataset dictionary (variables) for the datasets

\item \code{getData_history}: Obtains a data frame with dataset history containing publication date and version
}}

\examples{
 \dontrun{
 getData(7)
 getData("7")
 getData_dictionary(7)
 getData_history(7) }
}
\seealso{
\code{\link{availableDatasets}} to get the data frame with the list of available datasets and corresponding \code{id}
}
\author{
Alisher Suyunov
}
