% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDividends.R
\name{getDividends}
\alias{getDividends}
\title{Downloads a calendar of dividends}
\usage{
getDividends(stockType = c("", "privileged", "simple", "bond"))
}
\value{
Returns a data frame
}
\description{
Returns a data frame with the calendar of dividends from the Central Securities Depository (\url{uzcsd.uz}) database
}
\examples{
 \dontrun{
 getDividends()}
}
\author{
Alisher Suyunov
}
