% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMarketIndex.R
\name{getMarketIndex}
\alias{getMarketIndex}
\title{Downloading Market Indices (Uzbekistan Composite Index, UCI)}
\usage{
getMarketIndex(
  sector = c("all", "finance", "industry", "agriculture", "construction", "social",
    "transport", "trade", "other"),
  from = "01.01.2019",
  to = "dd.mm.yyyy"
)
}
\arguments{
\item{sector}{A string that indicates the sector: all, finance, industry, agriculture, construction, social, transport, trade, other}

\item{from}{A string representing the start date of the period of interest in date.month.year date format}

\item{to}{A string representing the end date of the period of interest in date.month.year date format}
}
\value{
Returns a data frame
}
\description{
Provides stock market indices over the period given
}
\examples{
 \dontrun{
 getMarketIndex(sector = "all")
 getMarketIndex(sector = "finance", from = "01.01.2020", to = "01.05.2020")}
}
\seealso{
\code{\link{getSecurities}} or \code{\link{RegisteredSecurities}} to see the complete list of available securities

\code{\link{getTicker}} to download stock prices of the certain stock
}
\author{
Alisher Suyunov
}
