% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecurities.R
\name{getSecurities}
\alias{getSecurities}
\title{Downloading the securities listed on the Republican Stock Exchange "Toshkent"}
\usage{
getSecurities()
}
\value{
Returns a data frame
}
\description{
Returns a data frame with the list of securities in the Republican Stock Exchange "Toshkent" database
}
\examples{
 \dontrun{
 getSecurities()}
}
\author{
Alisher Suyunov
}
