% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTicker.R
\name{getTicker}
\alias{getTicker}
\title{Downloading Stock Market Data}
\usage{
getTicker(symbol, from = "01.01.2020", to = "dd.mm.yyyy")
}
\arguments{
\item{symbol}{A unique identifier of the security (e.g. UZ7011340005). Note: Don't confuse with ISIN code}

\item{from}{A string representing the start date of the period of interest in date.month.year date format}

\item{to}{A string representing the end date of the period of interest in date.month.year date format}
}
\value{
Returns a data frame
}
\description{
Obtains stock market prices over the period given
}
\examples{
 \dontrun{
 getTicker("UZ7011340005")
 getTicker("UZ7011340005", from = "01.01.2020", to = "01.05.2020")}
}
\seealso{
\code{\link{getSecurities}} or \code{\link{RegisteredSecurities}} to see the complete list of available securities

\code{\link{getMarketIndex}} to download stock market indices by sector
}
\author{
Alisher Suyunov
}
