% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipo.R
\name{ipo}
\alias{ipo}
\title{Obtaining schedule of public offerings (IPO/SPO/PO (the government shares))}
\usage{
ipo(key = "", plus_n_years = 1)
}
\arguments{
\item{key}{A unique identifier of the security (e.g. UZ7011340005). Note: Don't confuse with ISIN code}

\item{plus_n_years}{End of the period}
}
\value{
Returns a data frame
}
\description{
Returns a data frame with public offerings in last five years
}
\examples{
 \dontrun{
 ipo()
 ipo(plus_n_years = 2)}
}
\author{
Alisher Suyunov
}
