% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_by_id.R
\name{lookup_by_id}
\alias{lookup_by_id}
\title{Looking up the dataset name by dataset id}
\usage{
lookup_by_id(id)
}
\arguments{
\item{id}{A numeric or string that represents the dataset identifier}
}
\value{
Returns a character
}
\description{
Provides the dataset's name that matches with dataset id
}
\examples{
 \dontrun{
 lookup_by_id(7)}
}
\author{
Alisher Suyunov
}
