% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegisteredSecurities.R
\name{RegisteredSecurities}
\alias{RegisteredSecurities}
\title{Downloads a full list of registered securities with ISIN and CFI codes}
\usage{
RegisteredSecurities()
}
\value{
Returns a data frame
}
\description{
Returns a data frame with the list of securities from the Central Securities Depository (\url{uzcsd.uz}) database
}
\examples{
 \dontrun{
 RegisteredSecurities()}
}
\author{
Alisher Suyunov
}
