% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentBidsAsks.R
\name{currentBidsAsks}
\alias{currentBidsAsks}
\title{Obtaining current Bids and Asks for stocks}
\usage{
currentBidsAsks(security_code = "", security_type = "STK")
}
\arguments{
\item{security_code}{An unique identifier of the security (e.g. UZ7011340005 or HMKB). Note: Don't confuse with ISIN code}

\item{security_type}{A string that indicates the type of the security: "STK" for stocks, "BND" for bonds, "RPO" for Repeat Public Offering, "FCT" for}
}
\value{
Returns a data frame
}
\description{
Returns current bid and ask prices on the Republican Stock Exchange "Toshkent"
}
\examples{
 \dontrun{
 currentBidsAsks()}
}
\seealso{
\code{\link{getSecurities}} or \code{\link{RegisteredSecurities}} to see the complete list of available securities

\code{\link{getTicker}} to download stock prices of the certain stock
}
\author{
Alisher Suyunov
}
